Move archiving utility functions to utils
